projects/aws-sms-monitor-1/api/requirements/prod.txt
Flask==1.0.2
redis==2.10.6
boto3==1.12.46
twilio==6.38.1
flask_script==2.0.6
SQLAlchemy==1.3.17
flask-debugtoolbar==0.11.0
python-dotenv==0.13.0
